game_fill  opraveny chyby pri ukládání v sezone 2018

+++ src/main/java/com/example/demo/Services/PlayerService.java
@@ -5,8 +5,8 @@ import com.example.demo.model.Player;
 import java.util.List;
 
 public interface PlayerService {
-    String playerNameById(Integer playerID);
-    Long playerTeamById(Integer playerID, String season);
+    String playerNameById(Long playerID);
+    Long playerTeamById(Long playerID, String season);
     Player lastPayerAdded();
     List <String> playersTeamsBySeasons(Player player);
     String[][] last5PlayersTeamsBySeasons();
